% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif-plot.R
\name{ccmotif.barplot2}
\alias{ccmotif.barplot2}
\title{Bar diagram with circular codes motif length distribution.}
\usage{
ccmotif.barplot2(ml, codeid = "unkn. code")
}
\arguments{
\item{codeid}{}
}
\description{
Bar diagram with circular codes motif length distribution.
}
