% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif-plot.R
\name{ccmotif.barplotDiff}
\alias{ccmotif.barplotDiff}
\title{Bar diagram with circular codes motif length distribution.}
\usage{
ccmotif.barplotDiff(ml, tml, codeid = "unkn. code")
}
\arguments{
\item{ml}{Classes for motif lengths}

\item{tml}{Classes for theoretical distribution of motif lenghts.}

\item{codeid}{}
}
\description{
Bar diagram with circular codes motif length distribution.
}
