% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds.R
\name{ccmotif.checkCDS.fasta}
\alias{ccmotif.checkCDS.fasta}
\title{Check a fasta file entry for structure of coding sequence.}
\usage{
ccmotif.checkCDS.fasta(ff, source, logfilename = NULL)
}
\arguments{
\item{ff}{Fasta file entry returned by `read.fasta`}

\item{source}{The source location of the fasta file.}

\item{logfilename}{File name for the logging data.}
}
\description{
Check a fasta file entry for structure of coding sequence.
}
