% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{ccmotif.classes}
\alias{ccmotif.classes}
\title{Create histogram classes for motifs (runs).}
\usage{
ccmotif.classes(ml, p, K = 8)
}
\arguments{
\item{ml}{List of motif lengths}

\item{p}{Probability (relative frequency) of code usage}
}
\value{
List of two elements: 1) sample histogram,
2) geometric distribution histogram
}
\description{
Create histogram classes for motifs (runs).
}
