% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{ccmotif.lengths.fasta}
\alias{ccmotif.lengths.fasta}
\title{Motif lengths in sequences of a FASTA file.}
\usage{
ccmotif.lengths.fasta(seqlist, code, frame = 0)
}
\arguments{
\item{seqlist}{Fasta file set (as returned by seqinr fasta.read)}

\item{frame}{Frame to use (0, 1 or 2). Default: 0}
}
\description{
Motif lengths in sequences of a FASTA file.
}
