% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{ccmotif.lengths}
\alias{ccmotif.lengths}
\title{Motif lengths in a sequence.}
\usage{
ccmotif.lengths(seq, code)
}
\arguments{
\item{seq}{Sequence}

\item{code}{Code as a list with entries id, and codons,
e.g. list(id = "foo", codons = c("ACU", "GAG")).}
}
\value{
Data structure with two elements: 1) incode is a vector of
motifs (run lengths) and 2) outcode is the vector of run lengths which are
not part of the code.
}
\description{
Motif lengths in a sequence.
}
