% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif-plot.R
\name{ccmotif.motif.hist}
\alias{ccmotif.motif.hist}
\title{Plot histogram for motif lengths in all reading frames.}
\usage{
ccmotif.motif.hist(ml0, ml1, ml2, species = "unkn.", minCu = 1,
  maxCu = 2)
}
\arguments{
\item{species}{A labe for a species.}

\item{minCu}{Left interval border.}

\item{maxCu}{Right interval border.}

\item{cu0}{Code usage for reading frame 0.}

\item{cu1}{Code usage for reading frame 1.}

\item{cu2}{Code usage for reading frame 2.}
}
\description{
Plot histogram for motif lengths in all reading frames.
}
