% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{ccmotif.nextLength}
\alias{ccmotif.nextLength}
\title{Next motif length.}
\usage{
ccmotif.nextLength(seq, code, max = 25, start = 1, fwd = 1)
}
\arguments{
\item{seq}{Sequence}

\item{code}{Code as a list with entries id, and codons,
e.g. list(id = "foo", codons = c("ACU", "GAG")).}

\item{max}{Maximal length. Default: 25.}
}
\description{
+ 1 2 3
1 | 1 2 3
2 | 1 3 2
3 | 1 2 3
}
