% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{ccmotif.readCodes}
\alias{ccmotif.readCodes}
\title{Read a code from file.
The structure of the file is:
# header with short description
1, ACA, ACT, ...
2, ACT, ATT, ...}
\usage{
ccmotif.readCodes(filename, tupleSize = 3, skipCodeIDColumn = TRUE)
}
\arguments{
\item{filename}{}

\item{tupleSize}{The size of the tuples in the code. Default is 3.}

\item{skipCodeIDColumn}{If true the first column is left out.}
}
\description{
Read a code from file.
The structure of the file is:
# header with short description
1, ACA, ACT, ...
2, ACT, ATT, ...
}
