% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{ccmotif.readMotifLengths}
\alias{ccmotif.readMotifLengths}
\title{Read a motif lengths file into a data frame.}
\usage{
ccmotif.readMotifLengths(s, results_path_prefix, frame = 0,
  rnd = FALSE, motif = TRUE, newZipVersion = FALSE)
}
\arguments{
\item{s}{Species descriptor}

\item{results_path_prefix}{Prefix for building the the path to
the results directory.}

\item{frame}{Which frame (0, 1 or 2). Default is 0.}

\item{rnd}{If true, motif lengths for random codes are read,
otherwise normal ones. Default is FALSE.}

\item{motif}{If true motif sequences are read, otherwise non-motif ones.}
}
\value{
Motif lengths as a data frame.
}
\description{
Read a motif lengths file into a data frame.
}
