% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{ccmotif.scan.fasta}
\alias{ccmotif.scan.fasta}
\title{Analysis of motif lengths in a list of fasta sequences.}
\usage{
ccmotif.scan.fasta(ff, codes, frame = 0, label = "unkn",
  tmpDir = "./")
}
\arguments{
\item{ff}{Entries of a fasta file set.}

\item{codes}{List of codes to analyze.}

\item{frame}{Frame to analyze (0, 1 or 2). Default is 0.}

\item{tmpDir}{Path to directory to save files. Default: "./".}

\item{seqid}{A short label (one word) for the sequences.}
}
\description{
The function scans over all DNA (RNA) sequences (\code{seqlist}) and
calculates the motif lengths for all codes (\code{codes}). Depending
on the parameters this might take a while.
This scan writes two comma separated files:
1) table with relative frequency of codons usage. Filename is
\code{code_usage_<seqid>.csv}
2) two files with the motif lengths of 1) the code and 2) codons not
in the code. Filename are
\code{motif_lengths_<seqid>.csv} and \code{motif_lengths_non_<seqid>.csv}
}
