% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{ccmotif.seq2cc}
\alias{ccmotif.seq2cc}
\title{Map a sequence to classes "in code" (1) or "out of code" (0).}
\usage{
ccmotif.seq2cc(seq, code)
}
\arguments{
\item{seq}{Sequence}

\item{code}{Code as a list with entries id, and tuples,
e.g. list(id = "foo", codons = c("ACU", "GAG")).}
}
\value{
Vector of 0s and 1s.
}
\description{
Map a sequence to classes "in code" (1) or "out of code" (0).
}
