% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{ccmotif.writeCodes}
\alias{ccmotif.writeCodes}
\title{Write codes to a file.}
\usage{
ccmotif.writeCodes(filename, codes)
}
\arguments{
\item{filename}{}

\item{codes}{List of codes.}
}
\description{
Write codes to a file.
}
