% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{codes.c3.equiv}
\alias{codes.c3.equiv}
\title{Equivalence class for a C3 code number.}
\usage{
codes.c3.equiv(cid)
}
\arguments{
\item{cid}{C3 code number.}
}
\value{
Its equivalence class.
}
\description{
Equivalence class for a C3 code number.
}
