% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\docType{data}
\name{codes.c3.equivmatrix}
\alias{codes.c3.equivmatrix}
\title{Table for mapping of code numbers to equivalence classes.}
\format{An object of class \code{matrix} with 216 rows and 2 columns.}
\usage{
codes.c3.equivmatrix
}
\value{
First column: C3 code number, second row: equivalence class.
}
\description{
Table for mapping of code numbers to equivalence classes.
}
\keyword{datasets}
