% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{codes.c3.inclass}
\alias{codes.c3.inclass}
\title{All code numbers for given equivalance class}
\usage{
codes.c3.inclass(eid)
}
\arguments{
\item{eid}{Equivalance class.}
}
\value{
List of C3 codes numbers.
}
\description{
All code numbers for given equivalance class
}
