% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\docType{data}
\name{codes.c3}
\alias{codes.c3}
\title{List of all 216 self complementary circular C3 codes.}
\format{An object of class \code{list} of length 216.}
\usage{
codes.c3
}
\description{
List of all 216 self complementary circular C3 codes.
}
\keyword{datasets}
