% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{codes.complByC3Id}
\alias{codes.complByC3Id}
\title{Shows the differences of two C3 codes addressed by its id.}
\usage{
codes.complByC3Id(i1, i2)
}
\arguments{
\item{i1}{Id of first code (1 to 216)}

\item{i2}{Id of second code (1 to 216)}
}
\value{
List of different codons.
}
\description{
Shows the differences of two C3 codes addressed by its id.
}
