% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{codes.intersectByC3Id}
\alias{codes.intersectByC3Id}
\title{Calculates the intersection of two C3 codes addressed by its id.}
\usage{
codes.intersectByC3Id(i1, i2)
}
\arguments{
\item{i1}{Id of first code (1 to 216)}

\item{i2}{Id of second code (1 to 216)}
}
\value{
List of shared codons.
}
\description{
Calculates the intersection of two C3 codes addressed by its id.
}
