% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{codes.random}
\alias{codes.random}
\title{Random code.}
\usage{
codes.random(size = 20, tuplelength = 3, id = "unkn. rnd",
  uracil = FALSE)
}
\arguments{
\item{size}{Number of codons in the code. Default is 20.}

\item{tuplelength}{Number of bases per tuple. Default is 3.}

\item{id}{Brief description of the code. Default is "unkn. rnd.".}

\item{uracil}{If false, all Us are replaced by Ts.}
}
\value{
Vector or string based codons.
}
\description{
Random code.
}
