% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{codes.readCodeUsage}
\alias{codes.readCodeUsage}
\title{Read a code usage file into a data frame.}
\usage{
codes.readCodeUsage(s, results_path_prefix, frame = 0, rnd = FALSE)
}
\arguments{
\item{s}{Species descriptor}

\item{results_path_prefix}{Prefix for building the the path to
the results directory.}

\item{frame}{Which frame (0, 1 or 2). Default is 0.}

\item{rnd}{If true random codes are read. Otherwise C3 codes.}
}
\value{
Code usage as a data frame.
}
\description{
Read a code usage file into a data frame.
}
