% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{codes.usage.hist}
\alias{codes.usage.hist}
\title{Plot histogram for code usage in all reading frames.}
\usage{
codes.usage.hist(cu0, cu1, cu2, species = "unkn.", minCu = 0.125,
  maxCu = 0.5)
}
\arguments{
\item{cu0}{Code usage for reading frame 0.}

\item{cu1}{Code usage for reading frame 1.}

\item{cu2}{Code usage for reading frame 2.}

\item{species}{A labe for a species.}

\item{minCu}{Left interval border.}

\item{maxCu}{Right interval border.}
}
\description{
Plot histogram for code usage in all reading frames.
}
