% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{codes.usage}
\alias{codes.usage}
\title{Calculates the code usage.}
\usage{
codes.usage(cu, code)
}
\arguments{
\item{cu}{Data frame for codon usages.}

\item{code}{A code.}
}
\value{
Percentage (0 to 1) of code usage.
}
\description{
Calculates the code usage.
}
