% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon.R
\name{codon.split}
\alias{codon.split}
\title{Split string sequence into tupels (codons).}
\usage{
codon.split(seq, size = 3)
}
\arguments{
\item{seq}{Sequence as a string.}

\item{size}{Tuple size. Default: 3 for codons.}
}
\value{
Vector of tuples. Each tuple is a string.
}
\description{
If the last tuple has less than `size` bases it is skipped.
}
