% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon.R
\name{codon.splitlist}
\alias{codon.splitlist}
\title{Split string sequences in a list into tupels (codons).}
\usage{
codon.splitlist(seqlist, size = 3)
}
\arguments{
\item{seqlist}{List of sequences (e.g. from a fasta file).}
}
\value{
Vector of tuples. Each tuple is a string.
}
\description{
The bases in every list are converted into tuples and
put together into a list of tuples (unflatten).
}
