% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon.R
\name{codon.usage}
\alias{codon.usage}
\title{Frequencies of codon usage.}
\usage{
codon.usage(codSeq)
}
\arguments{
\item{codSeq}{Sequence of tuples (codons)}
}
\value{
data frame with two columns:
1. codon, 2. number of codons in sequence
}
\description{
Calculates the absolute frequencies of codons in a codon sequence.
}
